Use double (not float) for Buffer2D type; removed unused Buffer3D type (#72)
* Use double instead of float for Buffer2D type; removed unused Buffer3D type

* Updated casts

+++ src/CAghostnodes.hpp
@@ -11,8 +11,8 @@
 #include <Kokkos_Core.hpp>
 
 // Load data (GrainID, DOCenter, DiagonalLength) into ghost nodes if the given RankY is associated with a 1D halo region
-KOKKOS_INLINE_FUNCTION void loadghostnodes(const float GhostGID, const float GhostDOCX, const float GhostDOCY,
-                                           const float GhostDOCZ, const float GhostDL, const int BufSizeX,
+KOKKOS_INLINE_FUNCTION void loadghostnodes(const double GhostGID, const double GhostDOCX, const double GhostDOCY,
+                                           const double GhostDOCZ, const double GhostDL, const int BufSizeX,
                                            const int MyYSlices, const int RankX, const int RankY, const int RankZ,
                                            const bool AtNorthBoundary, const bool AtSouthBoundary,
                                            Buffer2D BufferSouthSend, Buffer2D BufferNorthSend) {
@@ -38,8 +38,8 @@ KOKKOS_INLINE_FUNCTION void loadghostnodes(const float GhostGID, const float Gho
 // Load data (GrainID, DOCenter, DiagonalLength) into ghost nodes if the given RankX and RankY is associated with a 2D
 // halo region
 KOKKOS_INLINE_FUNCTION void
-loadghostnodes(const float GhostGID, const float GhostDOCX, const float GhostDOCY, const float GhostDOCZ,
-               const float GhostDL, const int BufSizeX, const int BufSizeY, const int MyXSlices, const int MyYSlices,
+loadghostnodes(const double GhostGID, const double GhostDOCX, const double GhostDOCY, const double GhostDOCZ,
+               const double GhostDL, const int BufSizeX, const int BufSizeY, const int MyXSlices, const int MyYSlices,
                const int RankX, const int RankY, const int RankZ, const bool AtNorthBoundary,
                const bool AtSouthBoundary, const bool AtWestBoundary, const bool AtEastBoundary,
                Buffer2D BufferSouthSend, Buffer2D BufferNorthSend, Buffer2D BufferWestSend, Buffer2D BufferEastSend,
